refactor: ga 파일 shared로 경로 변경
